+++ ospatch/updates_linux.go
@@ -20,7 +20,6 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"syscall"
 
 	osconfigpb "github.com/GoogleCloudPlatform/osconfig/_internal/gapi-cloud-osconfig-go/google.golang.org/genproto/googleapis/cloud/osconfig/v1alpha1"
 	"github.com/GoogleCloudPlatform/osconfig/inventory/packages"
@@ -91,25 +90,4 @@ func runUpdates(r *patchRun) error {
 		return err
 	}
 	return packages.UpdatePackages()
-}
-
-func rebootSystem() error {
-	// Start with systemctl and work down a list of reboot methods.
-	if e, _ := exists(systemctl); e {
-		logger.Debugf("Rebooting using systemctl.")
-		return exec.Command(systemctl, "reboot").Run()
-	}
-	if e, _ := exists(reboot); e {
-		logger.Debugf("Rebooting using reboot command.")
-		return exec.Command(reboot).Run()
-	}
-	if e, _ := exists(shutdown); e {
-		logger.Debugf("Rebooting using shutdown command.")
-		return exec.Command(shutdown, "-r", "-t", "0").Run()
-	}
-
-	// Fall back to reboot(2) system call
-	logger.Debugf("No suitable reboot command found, rebooting using reboot(2).")
-	syscall.Sync()
-	return syscall.Reboot(syscall.LINUX_REBOOT_CMD_RESTART)
 }
